fix: secret key only uses byte[]
Break inheritance with Key superclass as the byte[] is all we need.

All the time the key is 32 of length and by using a type for SecretKey we don't need to check as the code does not allow to pass a SecretKey instance to a PrivateKey instance. Typing to the rescue.