refactor: remove private/public key from commons and use typed keys expected by the algorithms (#243)
Break inheritance with Key superclass as the byte[] is all we need.

The secret key is 32 of length and by using a type for SecretKey we don't need to check as the code does not allow to pass a SecretKey instance to a PrivateKey instance. Typing to the rescue.

Public/private key define own types in each version.